Select policy by short name and further refactorings.
